@@ -14,7 +14,7 @@
 		}
 	],
 	"require": {
-		"php": ">=7.0.9",
+		"php": ">=7.4",
 		"laravel/framework": "^6.20.26|^7.0|^8.0|^9.0"
 	},
 	"require-dev": {
